requirements.txt
Flask==0.12.2
pydns==2.3.6
uWSGI==2.0.15
